+++ CHANGELOG.md
@@ -1,9 +1,12 @@
+## 0.0.30
+
+* Add support for `encoding` parameter
+
 ## 0.0.29
 
 * Support .msg files
 * Refactor parallel mode and add smoke test
 * Fix header value for api key
-* Add support for `encoding` parameter
 
 ## 0.0.28